requirements.txt
apache-airflow
s3fs
boto3
pandas
numpy
mlflow
scikit-learn
imbalanced-learn
